split identidt into small files and create seperate test file for each one
